move: 덱상세페이지 패키지 수정
